Add Override file for unkn/717 display
This is the display I received with my OSCAR kit, but it seems not all OSCARs
ship with exactly the same panel. Some, at least, have the Qualia 610/c111 IDs.

It turns out my build of Yosemite already has a file for 756e6b6e/717, so I
added the meaty bits from the Qualia file to the shipped one. Works for me.

Signed-off-by: Chris Boot <bootc@bootc.net>